requirements.txt
streamlit>=1.36.0
python-dotenv>=1.0.1
langchain>=0.2.12
langchain-openai>=0.1.21
langgraph>=0.2.2
langsmith>=0.1.0
openai>=1.0.0
faiss-cpu==1.8.0.post1
requests>=2.31.0
beautifulsoup4>=4.12.0
yt-dlp>=2024.3.10
plotly>=5.17.0
pandas>=2.0.0